Avance 06
Correccion Registro Usuario y Asignación Rol

+++ Proyecto/app/Http/Controllers/PermissionController.php
@@ -3,17 +3,19 @@
 namespace App\Http\Controllers;
 
 use App\Permission;
-
 use App\Role;
-
 use App\Http\Requests\Permission\StoreRequest;
-
 use App\Http\Requests\Permission\UpdateRequest;
-
 use Illuminate\Http\Request;
 
 class PermissionController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('role:' . config('app.admin_role'));
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -21,8 +23,9 @@ class PermissionController extends Controller
      */
     public function index()
     {
-        return view('theme.backoffice.pages.permission.index', [
-            'permissions' => Permission::all(),
+        $this->authorize('index',Role::class);
+        return view('theme.backoffice.pages.permission.index',[
+            'permissions' => Permission::all()
         ]);
     }
 
@@ -33,6 +36,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        $this->authorize('create',Role::class);
         return view('theme.backoffice.pages.permission.create',[
             'roles' => Role::all()
         ]);
@@ -47,7 +51,6 @@ class PermissionController extends Controller
     public function store(StoreRequest $request, Permission $permission)
     {
         $permission = $permission->store($request);
-        alert()->success('Guardado', 'Exito')->autoclose(2000);
         return redirect()->route('backoffice.permission.show', $permission);
     }
 
@@ -59,8 +62,9 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
-        return view('theme.backoffice.pages.permission.show', [
-            'permission' => $permission,
+        $this->authorize('view', $permission);
+        return view('theme.backoffice.pages.permission.show',[
+            'permission' => $permission
         ]);
     }
 
@@ -72,9 +76,10 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
-        return view('theme.backoffice.pages.permission.edit', [
+        $this->authorize('update', $permission);
+        return view('theme.backoffice.pages.permission.edit',[
             'permission' => $permission,
-            'roles' => Role::all(),
+            'roles' => Role::all()
         ]);
     }
 
@@ -88,8 +93,7 @@ class PermissionController extends Controller
     public function update(UpdateRequest $request, Permission $permission)
     {
         $permission->my_update($request);
-        alert()->success('Actualizado', 'Exito')->autoclose(2000);
-        return redirect()->route('backoffice.permission.show', $permission);
+        return redirect()->route('backoffice.permission.show',$permission);
     }
 
     /**
@@ -100,10 +104,10 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
+        $this->authorize('delete', $permission);
         $role = $permission->role;
         $permission->delete();
         alert()->success('     ', 'Permiso eliminado')->autoclose(2000);
-        //return redirect()->route('backoffice.permission.index');
         return redirect()->route('backoffice.role.show', $role);
     }
 }
