+++ Proyecto/app/Http/Controllers/RoleController.php
@@ -3,85 +3,55 @@
 namespace App\Http\Controllers;
 
 use App\Role;
-
 use App\Http\Requests\Role\StoreRequest;
-
 use App\Http\Requests\Role\UpdateRequest;
-
 use Illuminate\Http\Request;
 
 class RoleController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    public function __construct()
+    {
+        $this->middleware('role:' . config('app.admin_role'));
+    }
+    
     public function index()
     {
-        //dd(Role::all());
-        return view('theme.backoffice.pages.role.index',[
+        $this->authorize('index', Role::class);
+        return view('theme.backoffice.pages.role.index', [
             'roles' => Role::all(),
         ]);
     }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    
     public function create()
     {
+        $this->authorize('create',Role::class);
         return view('theme.backoffice.pages.role.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(StoreRequest $request,Role $role)
+    public function store(StoreRequest $request, Role $role)
     {
         $role = $role->store($request);
         alert()->success('Guardado', 'Exito')->autoclose(2000);
-        return redirect()->route('backoffice.role.show',$role);
+        return redirect()->route('backoffice.role.show', $role);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Role  $role
-     * @return \Illuminate\Http\Response
-     */
     public function show(Role $role)
     {
+        $this->authorize('view',$role);
         return view('theme.backoffice.pages.role.show', [
             'role' => $role,
             'permissions' => $role->permissions
         ]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Role  $role
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Role $role)
     {
+        $this->authorize('update', $role);
         return view('theme.backoffice.pages.role.edit', [
             'role' => $role,
         ]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Role  $role
-     * @return \Illuminate\Http\Response
-     */
     public function update(UpdateRequest $request, Role $role)
     {
         $role->my_update($request);
@@ -89,14 +59,9 @@ class RoleController extends Controller
         return redirect()->route('backoffice.role.show', $role);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Role  $role
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(Role $role)
     {
+        $this->authorize('delete', $role);
         $role->delete();
         alert()->success('     ', 'Rol eliminado')->autoclose(2000);
         return redirect()->route('backoffice.role.index');
